+++ _icon/stellaris_checksum_patcher_icon.ico
