+++ src/UI/StellarisChecksumPatcherUI.py
