+++ src/UI/StellarisChecksumPatcherUI.ui
