+++ src/UI/ui_icons/patch_icon.png
