+++ src/UI/ui_icons/stellaris_checksum_patcher_icon.ico
