+++ src/patchers/HexPatcher-deprecated.py
