+++ src/patchers/save_patcher.py
@@ -8,7 +8,7 @@ import zipfile
 import tempfile
 
 # 3rd Party
-from src.utils.global_defines import logger, settings, OS
+from utils.global_defines import logger, settings, OS
 
 Path = pathlib.Path
